+++ src/Logging.ts
@@ -4,7 +4,6 @@ export type ErrorLogger = (options: ErrorLoggerOptions) => void;
 
 export interface ErrorLoggerOptions {
   //options: Required<ErrorOptions>;
-  source: string;
   message: string;
   error: SendableError;
   info: any;
@@ -13,18 +12,15 @@ export interface ErrorLoggerOptions {
 }
 
 export const defaultErrorLogger: ErrorLogger = (options) => {
-  const { source, message, error, info } = options;
+  const { message, error, info } = options;
 
   const stackLines = error.stack?.split("\n").slice(1).join("\n");
-  console.error(
-    `${source || "SendableError"} - ${message} ${JSON.stringify(info || {})}${stackLines ? `\n${stackLines}` : ""}`,
-  );
+  console.error(`SendableError - ${message} ${JSON.stringify(info || {})}${stackLines ? `\n${stackLines}` : ""}`);
 
   const cause = error.cause;
   if (cause && cause instanceof Error) {
     defaultErrorLogger({
       ...options,
-      source,
       message: `Caused by: ${cause.message}`,
       error: cause as SendableError,
       errorInfo: {},